backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
qdrant-client
pydantic